requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
factory-boy==3.2.1
Faker==13.13.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.6.0
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
